% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sc_varipath.R
\name{plot_sc_varipath}
\alias{plot_sc_varipath}
\title{plot_sc_varipath()}
\usage{
plot_sc_varipath(sc_varipath_out, type = NULL)
}
\arguments{
\item{sc_varipath_out}{The output list from the sc_varipath() function.}

\item{type}{Type of plot to generate. Options are 'perturbation_to_RC', 'RC_to_gene', or 'RC_to_pathway'.}
}
\value{
A dotplot with dendrograms showing associations between perturbations and rotated components, rotated components and genes, or rotated components and pathways.
}
\description{
Plot results from sc_varipath() function to identify pathways associated with perturbations.
}
\details{
This function visualizes the relationships identified by the sc_varipath() function. Depending on the specified type, it generates a dotplot with dendrograms to illustrate the associations between perturbations and rotated components, rotated components and genes, or rotated components and pathways. The size and color of the dots represent the strength of the associations.
}
\examples{
out = sc_varipath(seurat_obj = adata.R, perturbation_column = 'miR.family')

options(repr.plot.width=20, repr.plot.height=10)
plot_sc_varipath(sc_varipath_out = out, type = 'perturbation_to_RC')
}
