% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_varipath.R
\name{sc_varipath}
\alias{sc_varipath}
\title{sc_varipath()}
\usage{
sc_varipath(
  seurat_obj,
  perturbation_column = NULL,
  enrichR_dbs = "GO_Biological_Process_2021",
  loading_cutoff = 0.1,
  FDR = 0.25
)
}
\arguments{
\item{seurat_obj}{A Seurat object containing the single-cell RNA-seq data with PCA already run.}

\item{perturbation_column}{The name of the metadata column in the Seurat object containing cell perturbation/guide information.}

\item{enrichR_dbs}{A character vector of enrichR database names to use for pathway enrichment analysis. Default is 'GO_Biological_Process_2021'.}

\item{loading_cutoff}{A numeric value indicating the loading cutoff to use for selecting genes associated with each rotated component. Default is 0.1.}

\item{FDR}{A numeric value indicating the FDR threshold to use for selecting significant components based on guide association p-values. Default is 0.25.

It is advised that you save the output of this function as an RDS file using saveRDS() immediately after running it for easy loading in future sessions and because memory failure can occur when running downstream plotting functions on large datasets.}
}
\value{
A list with the following components:
\itemize{
\item{vari_weights}{Varimax weights of genes for each rotated component.}
\item{fitted_models}{Linear models fitting perturbation to each rotated component.}
\item{fit_pvals_adj}{Adjusted p-values of the association between perturbation and each rotated component to identify best fit for each perturbation.}
\item{perturbation_RC_df}{Adjusted p-values of the association between perturbation and each rotated component to plot using plot_sc_varipath().}
\item{weights_dataframe}{Rotated component gene loadings to be used for plot_sc_varipath().}
\item{adjacency_table}{Contains 3 most associated pathways for each rotated component to plot using plot_sc_varipath().}
}
}
\description{
Identify novel pathway shifts associated with perturbations in single-cell RNA-seq data using a varimax rotation of scRNA-seq PCA.
}
\examples{
out = sc_varipath(seurat_obj = adata.R, perturbation_column = 'miR.family')
}
